Reusing DelayBufferedOutputStream also for master-side writes to the log file.
